Refactor use of LatexView in the app
